Extract test base
